requirements.txt
# Core Flask
Flask==2.3.3
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1

# Database
PyMySQL==1.1.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21

# Authentication & Security
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0

# Forms & Validation
WTForms==3.0.1

# Date & Time
python-dateutil==2.8.2

# Development
Flask-DebugToolbar==0.13.1

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Optional: Email support
Flask-Mail==0.9.1

# Optional: API Rate Limiting
Flask-Limiter==3.5.0
redis==5.0.0

# Optional: Internationalization
Flask-Babel==3.1.0

# Optional: Caching
Flask-Caching==2.1.0

# Optional: File uploads
Pillow==10.0.1

# Testing
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2

# Logging
python-json-logger==2.0.7